revert back to normal functions
