--- netlify.toml
@@ -1,7 +1,0 @@
-[build]
-  command = "npm run build"
-  publish = ".next"
-
-[[edge_functions]]
-  function = "og"
-  path = "/og"
